% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.anderson_rubin.R
\name{my.anderson_rubin}
\alias{my.anderson_rubin}
\title{Anderson-Rubin CI for TSLS}
\usage{
my.anderson_rubin(
  Yvar,
  Xvar,
  Inc,
  Exc,
  data,
  intercept = TRUE,
  tolerance = 1e-16,
  beta_low,
  beta_high,
  init_length = 10,
  precision
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvar}{Endogenous regressor}

\item{Inc}{Included instruments}

\item{Exc}{Excluded instruments}

\item{data}{Of type data.table}

\item{intercept}{Logical, if including 1 as model intercept}

\item{tolerance}{Tolerance level for matrix inversion}

\item{beta_low}{Lower bound for search}

\item{beta_high}{Upper bound for search}

\item{init_length}{Size of search grid}

\item{precision}{Maximum difference between points in and out of CI}
}
\description{
This function returns point estimates and the variance-covariance matrix.
Written for data.table arguments.
}
\examples{
my.anderson_rubin()
}
\keyword{anderson-rubin}
\keyword{tsls}
