% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.jackknife_iv_boot.R
\name{my.jackknife_iv_boot}
\alias{my.jackknife_iv_boot}
\title{Jackknife IV with Bootstrapped SEs}
\usage{
my.jackknife_iv_boot(
  Yvar,
  Xvar,
  Inc,
  Exc,
  data,
  intercept = TRUE,
  tolerance = 1e-16,
  n_boot = 100
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvar}{Endogenous regressor}

\item{Inc}{Included instruments}

\item{Exc}{Excluded instruments}

\item{data}{Of type data.table}

\item{intercept}{Logical, if including 1 as model intercept}

\item{tolerance}{Tolerance level for matrix inversion}

\item{n_boot}{Number of bootstrapped samples}
}
\description{
This function returns point estimates and the variance-covariance matrix.
Written for data.table arguments.
}
\examples{
my.jackknife_iv_boot()
}
\keyword{bootstrap}
\keyword{iv}
\keyword{jackknife}
\keyword{linear}
