% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.knn.binary.R
\name{my.knn.binary}
\alias{my.knn.binary}
\title{K-nearest neighbors matching - binary treatment}
\usage{
my.knn.binary(
  Yvar,
  Xvars,
  read_data,
  metric = "mahalanobis",
  binary_treat_var,
  tolerance = 1e-16,
  K = 1L
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvars}{Character vector of control variables}

\item{read_data}{Of type data.table - source of data}

\item{metric}{Metric for comparison}

\item{binary_treat_var}{Binary treatment indicator}

\item{tolerance}{Tolerance level for matrix inversion}

\item{K}{number of neighbors}

\item{predict_data}{Of type data.table - points for prediction}
}
\description{
For multiple metrics - Euclidean and Mahalanobis
Written for data.table arguments.
}
\examples{
my.knn.binary()
}
\keyword{nearest}
\keyword{neighbors}
