% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.knn.R
\name{my.knn}
\alias{my.knn}
\title{K-nearest neighbors smoother}
\usage{
my.knn(
  Yvar,
  Xvars,
  read_data,
  predict_data,
  metric = "mahalanobis",
  tolerance = 1e-16,
  K = 1L
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvars}{Character vector of control variables}

\item{read_data}{Of type data.table - source of data}

\item{predict_data}{Of type data.table - points for prediction}

\item{metric}{Metric for comparison}

\item{tolerance}{Tolerance level for matrix inversion}

\item{K}{number of neighbors}
}
\description{
For multiple metrics - Euclidean and Mahalanobis
Written for data.table arguments.
}
\examples{
my.knn()
}
\keyword{nearest}
\keyword{neighbors}
