% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.larf.R
\name{my.larf}
\alias{my.larf}
\title{Local Average Response Function (Abadie, 2003)}
\usage{
my.larf(Yvar, Dvar, Zvar, Xvar1, Cont1, Xvar2, data, K = 2, tolerance = 1e-16)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Dvar}{Treatment variable}

\item{Zvar}{Instrument}

\item{Xvar1}{First-stage running variable (only 1 variable for now)}

\item{Cont1}{Controls for first stage (should include FEs or intercept)}

\item{Xvar2}{Second stage control variables}

\item{data}{Of type data.table}

\item{K}{Degree of polynomial for first stage Xvar1}

\item{tolerance}{Tolerance level for matrix inversion}
}
\description{
This function returns point estimates and the variance-covariance matrix.
Written for data.table arguments.
}
\examples{
my.larf()
}
\keyword{response}
\keyword{treatment}
