% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.local_poly.R
\name{my.local_poly}
\alias{my.local_poly}
\title{Local polynomial estimator}
\usage{
my.local_poly(
  Yvar,
  Xvar,
  read_data,
  predict_data,
  tolerance = 1e-16,
  degree = 1L,
  bandwidth = 1
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvar}{Running variable}

\item{read_data}{Of type data.table - data for reading}

\item{predict_data}{Of type data.table - Xvar points for predicting values}

\item{tolerance}{Tolerance level for matrix inversion}

\item{degree}{Degree of polynomial}

\item{bandwidth}{Bandwidth of data for Xvar}
}
\description{
This function returns point estimates and the variance-covariance matrix.
Written for data.table arguments.
}
\examples{
my.local_poly()
}
\keyword{linear}
\keyword{local}
