% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.ols.R
\name{my.ols}
\alias{my.ols}
\title{Ordinary Least Squares}
\usage{
my.ols(Yvar, Xvars, data, intercept = TRUE, varcov = "het", tolerance = 1e-16)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvars}{Character vector of control variables}

\item{data}{Of type data.table}

\item{intercept}{Logical, if including 1 as model intercept}

\item{varcov}{Method for variance-covariance estimation}

\item{tolerance}{Tolerance level for matrix inversion}
}
\description{
This function returns point estimates and the variance-covariance matrix.
Written for data.table arguments.
}
\examples{
my.ols()
}
\keyword{linear}
\keyword{ols}
