% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.sieve.R
\name{my.sieve}
\alias{my.sieve}
\title{Sieve Estimators}
\usage{
my.sieve(
  Yvar,
  Xvar,
  read_data,
  predict_data,
  basis = "standard",
  low.bound.Xvar = 0,
  high.bound.Xvar = 1,
  tolerance = 1e-16,
  K = 1L
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvar}{Character vector of control variables}

\item{read_data}{Of type data.table - points for prediction}

\item{basis}{Standard for use}

\item{low.bound.Xvar}{Lower bound of data}

\item{high.bound.Xvar}{Upper bound of data}

\item{tolerance}{Tolerance level for matrix inversion}

\item{K}{degrees of polynomial (or knots for linear spline basis)}
}
\description{
For multiple bases - Standard, Bernstein, and Linear Spline
Written for data.table arguments.
}
\examples{
my.sieve()
}
\keyword{polynomial}
\keyword{sieve}
