% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.tsls.R
\name{my.tsls}
\alias{my.tsls}
\title{Ordinary Least Squares}
\usage{
my.tsls(
  Yvar,
  Xvar,
  Inc,
  Exc,
  data,
  intercept = TRUE,
  varcov = "het",
  tolerance = 1e-16
)
}
\arguments{
\item{Yvar}{Dependent variable}

\item{Xvar}{Endogenous variable}

\item{Inc}{Included instruments (character vector of controls)}

\item{Exc}{Excluded instruments (character vector)}

\item{data}{Of type data.table}

\item{intercept}{Logical, if including 1 as model intercept}

\item{varcov}{Method for variance-covariance estimation}

\item{tolerance}{Tolerance level for matrix inversion}
}
\description{
This function returns point estimates and the variance-covariance matrix.
Written for data.table arguments.
}
\examples{
my.tsls()
}
\keyword{linear}
\keyword{ols}
