% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.variance.R
\name{my.variance}
\alias{my.variance}
\title{Variance estimator}
\usage{
my.variance(vars, data, tolerance = 1e-16)
}
\arguments{
\item{vars}{Variables of interest}

\item{data}{Of type data.table}

\item{tolerance}{Tolerance level for matrix inversion}
}
\description{
Variance-covariance matrix
}
\details{
Written for data.table arguments.
}
\examples{
my.variance()
}
\keyword{variance}
